refactor: split users driving layer into api and impl modules
Extract public BackendUser.toDto() mapper into users:be:driving:api
so other features can convert BackendUser to UserApiDto without
depending on users:be:driving:impl.
