+++ feat/projects/be/driving/impl/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 dependencies {
     implementation(project(":feat:projects:be:driving:contract"))
     implementation(project(":feat:projects:be:app:api"))
-    implementation(project(":feat:users:be:driving:impl"))
+    implementation(project(":feat:users:be:driving:api"))
     implementation(project(":feat:users:be:driving:contract"))
     implementation(project(":feat:users:be:model"))
     testImplementation(project(":lib:configuration:be:api"))
