+++ feat/users/be/driving/api/src/main/kotlin/cz/adamec/timotej/snag/users/be/driving/api/Mapper.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.users.be.driving.api
+
+import cz.adamec.timotej.snag.users.be.driving.contract.UserApiDto
+import cz.adamec.timotej.snag.users.be.model.BackendUser
+
+fun BackendUser.toDto() =
+    UserApiDto(
+        id = user.id.toString(),
+        entraId = user.entraId,
+        email = user.email,
+        role = user.role?.name,
+        updatedAt = user.updatedAt.value,
+    )
