+++ feat/users/be/driving/impl/src/main/kotlin/cz/adamec/timotej/snag/users/be/driving/impl/internal/Mapper.kt
@@ -14,20 +14,10 @@ package cz.adamec.timotej.snag.users.be.driving.impl.internal
 
 import cz.adamec.timotej.snag.lib.core.common.Timestamp
 import cz.adamec.timotej.snag.users.be.driving.contract.PutUserApiDto
-import cz.adamec.timotej.snag.users.be.driving.contract.UserApiDto
 import cz.adamec.timotej.snag.users.be.model.BackendUser
 import cz.adamec.timotej.snag.users.business.User
 import cz.adamec.timotej.snag.users.business.UserRole
 import kotlin.uuid.Uuid
-
-fun BackendUser.toDto() =
-    UserApiDto(
-        id = user.id.toString(),
-        entraId = user.entraId,
-        email = user.email,
-        role = user.role?.name,
-        updatedAt = user.updatedAt.value,
-    )
 
 internal fun PutUserApiDto.toModel(id: Uuid) =
     BackendUser(
