+++ settings.gradle.kts
@@ -195,6 +195,7 @@ include(":feat:users:be:app:api")
 include(":feat:users:be:app:impl")
 include(":feat:users:be:ports")
 include(":feat:users:be:driven:impl")
+include(":feat:users:be:driving:api")
 include(":feat:users:be:driving:contract")
 include(":feat:users:be:driving:impl")
 include(":feat:users:fe:model")
